requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.23
cryptography==41.0.7
python-dotenv==1.0.0
google-genai==0.1.0
pandas==2.1.3
plotly==5.17.0
